requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# AI/ML dependencies
openai==1.3.7
replicate==0.15.4
huggingface-hub==0.19.4
transformers==4.36.0
torch==2.1.1
Pillow==10.1.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
sqlite3

# Utility
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring
sentry-sdk[fastapi]==1.38.0 